refactor: Result 객체 디렉토리를 dto에서 domain 아래로 변경
